fix: create constantData
